pyproject.toml
[project]
name = "cobblestone"
version = "0.1.0"
description = "Day-ahead electricity price forecasting with QA pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "polars>=1.0.0",
    "pyarrow>=23.0.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
    "numpy>=2.0.0",
    "click>=8.0.0",
    "pytz>=2024.0",
    "lightgbm>=4.6.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "duckduckgo-search>=8.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qa_pipeline", "src/remediation_pipeline", "src/interpretation_agent"]

[project.scripts]
qa-pipeline = "qa_pipeline.cli:main"
remediation-pipeline = "remediation_pipeline.cli:main"
interpret = "interpretation_agent.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
